+++ pyproject.toml
@@ -227,8 +227,8 @@ module = [
     "neptune.common.storage.storage_utils",
     "neptune.internal.utils.utils",
     "neptune.internal.warnings",
-    "neptune.common.websockets.reconnecting_websocket",
-    "neptune.common.websockets.websocket_client_adapter",
+    "neptune.internal.websockets.reconnecting_websocket",
+    "neptune.internal.websockets.websocket_client_adapter",
     "neptune.exceptions",
     "neptune.handler",
     "neptune.integrations.python_logger",
