+++ src/neptune/internal/websockets/websocket_signals_background_job.py
@@ -30,11 +30,11 @@ from neptune.attributes.constants import (
     SIGNAL_TYPE_STOP,
     SYSTEM_FAILED_ATTRIBUTE_PATH,
 )
-from neptune.common.websockets.reconnecting_websocket import ReconnectingWebsocket
 from neptune.internal.background_job import BackgroundJob
 from neptune.internal.threading.daemon import Daemon
 from neptune.internal.utils import process_killer
 from neptune.internal.utils.logger import get_logger
+from neptune.internal.websockets.reconnecting_websocket import ReconnectingWebsocket
 from neptune.internal.websockets.websockets_factory import WebsocketsFactory
 
 if TYPE_CHECKING:
